Weather/sunrise-sunset/pyproject.toml
[project]
name = "sunrise-sunset"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "ansicolors>=1.1.8",
    "colors-py==0.2.2",
    "pillow==10.3.0",
    "scipy>=1.13.1",
    "suntime==1.3.2",
]
